Update README.md
ABR Coding Activity
This project is a single-page application that helps determine the best region for the new headquarters of the American Board of Radiology. The app fetches data from the coding challenge server API and displays it in a user-friendly way.

Features
Fetches data from the coding challenge server
Navigation bar with links to each page of the site
Home page displaying a list of each NOAAFisheriesRegion from the dataset, each showing the average calories and fat per serving for the fish in that region
A page for each of the regions, showing:
The name of the region
The average calories and fat per serving for the fish in that region
A list of fish from that region with:
The name of the fish
An image
Calories/fat per serving
Some descriptive text
Tech Stack
React
Usage
Clone the repository to your local machine
Run npm install to install the dependencies
Run npm start to start the development server
Requirements
A JavaScript front-end framework must be used (Aurelia, React, Angular, Vue, Svelte, etc.)
Please do not use a front-end component library for templating (Material UI, Bootstrap, Chakra, Bulma, etc.). We’d like to see how you work with plain HTML and CSS.
Please take no more than 4 hours to complete this assignment.
This will be judged on:
Functionality.
Code is “clean” and maintainable.
Markup is well-organized.
Stretch goal: It looks nice.
Authors
Nika Darab
Acknowledgments
The American Board of Radiology
Coding Challenge Server